+++ src/Enum.ts
@@ -14,10 +14,10 @@ import { constant, flow, pipe } from "fp-ts/function"
 import { Ord } from "fp-ts/Ord"
 import { Bounded } from "fp-ts/Bounded"
 import * as O from "fp-ts/Option"
-import Option = O.Option
+type Option<A> = O.Option<A>
 import { unsafeExpect as unsafeExpectO } from "./Option"
 import * as NEA from "fp-ts/NonEmptyArray"
-import NonEmptyArray = NEA.NonEmptyArray
+type NonEmptyArray<A> = NEA.NonEmptyArray<A>
 import * as A from "fp-ts/Array"
 import { dup, toFst, toSnd } from "./Tuple"
 import * as Map from "fp-ts/Map"
@@ -25,7 +25,7 @@ import { Eq } from "fp-ts/Eq"
 import * as Semigroup from "fp-ts/Semigroup"
 import { increment, decrement, add } from "./Number"
 import * as L from "./Lazy"
-import Lazy = L.Lazy
+type Lazy<A> = L.Lazy<A>
 
 /**
  * Typeclass for finite enumerations.