+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to semantic versioning.
 
 ## 0.19.0 (_Unreleased_)
 
+- Fix typings in presence of `isolatedModules`.
 - Bump minimum supported Node LTS to v20.
 
 ## 0.18.0 (2023-10-26)